refactor: 갤러리 페이지 TNB 변경
