+++ src/app/place/[id]/components/index.ts
@@ -1,3 +1,4 @@
 export { default as InfoSection } from './InfoSection';
 export { default as ReviewSection } from './ReviewSection';
 export { default as GallerySection } from './GallerySection';
+export { default as PlaceTopNavigationBar } from './PlaceTopNavigationBar';
