+++ src/app/place/[id]/gallery/components/index.ts
@@ -1,1 +1,2 @@
 export { default as GalleryModal } from './GalleryModal';
+export { default as GalleryTopNavigationBar } from './GalleryTopNavigationBar';
