+++ src/app/place/[id]/gallery/page.tsx
@@ -1,11 +1,10 @@
 'use client';
 
-import { GalleryModal } from './components';
+import { GalleryModal, GalleryTopNavigationBar } from './components';
 import { useGetImages } from '@/apis/image';
 import { useGetPlace } from '@/apis/place';
 import { useGetReviews } from '@/apis/review';
 import { Footer } from '@/app/place/components';
-import { TopNavigationBar } from '@/components';
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
 import { useEffect, useState } from 'react';
@@ -47,29 +46,8 @@ export default function Page({ params }: { params: { id: string } }) {
 
   return (
     <>
-      <TopNavigationBar
-        title={placeData?.name}
-        onBackClick={() => router.push(`/place/${params.id}`)}
-        rightNode={
-          <>
-            <Image
-              src="/assets/icons/28/Bookmark.svg"
-              alt="Bookmark"
-              width={28}
-              height={28}
-              onClick={() => console.log('북마크 클릭')}
-            />
-            <Image
-              src="/assets/icons/28/Share.svg"
-              alt="Share"
-              width={28}
-              height={28}
-              onClick={() => setIsGalleryModalOpen(true)}
-            />
-          </>
-        }
-      />
-      <div className="grid grid-cols-3 gap-1 pt-14">
+      <GalleryTopNavigationBar placeId={placeId} name={placeData?.name ?? ''} />
+      <div className="grid grid-cols-3 gap-1">
         {galleryImages.map(({ url }) => (
           <div
             key={url}
