+++ src/app/place/components/Footer/index.tsx
@@ -11,7 +11,7 @@ export default function Footer() {
       <footer>
         <Button
           type="DEFAULT"
-          className="fixed inset-x-0 bottom-0 z-50 mx-auto w-full min-w-[360px] max-w-[448px]"
+          className="fixed inset-x-0 bottom-0 z-30 mx-auto w-full min-w-[360px] max-w-[448px]"
           onClick={() => setIsLogTimeSheetOpen(true)}
         >
           카공 기록하기
