+++ src/app/place/components/index.ts
@@ -3,4 +3,3 @@ export { default as ReviewItem } from './ReviewItem';
 export { default as ReviewSheet } from './ReviewSheet';
 export { default as TimeLogSheet } from './TimeLogSheet';
 export { default as ShareSheet } from './ShareSheet';
-export { default as PlaceTopNavigationBar } from './PlaceTopNavigationBar';
